japanese_media_manager/requirements.txt
beautifulsoup4==4.10.0
mergedeep==1.3.4
Pillow==9.0.1
PyYAML==6.0
requests==2.25.1
